pyproject.toml
[project]
name = "keyoku-demo"
version = "0.1.0"
description = "Interactive demo showcasing Keyoku AI Memory SDK"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.1.0",
    "python-dotenv>=1.0.0",
    "gradio>=4.0.0,<5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
keyoku-demo = "keyoku_demo.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/keyoku_demo"]

[tool.ruff]
line-length = 100
target-version = "py39"
